approch2/requirement.txt
fastapi
uvicorn[standard]
pdf2image
Pillow
pytesseract
PyMuPDF
requests
ollama
pydantic
